feat : add ROS2 Message Classes

+++ src/main/kotlin/net/wavem/uvc/ros2/builtin_interfaces/msg/Time.kt
@@ -1,7 +1,7 @@
 package net.wavem.uvc.ros2.builtin_interfaces.msg
 
 data class Time(
-    val sec: Float,
-    val nanosec: Float
+    val sec: Int,
+    val nanosec: Int
 ) {
 }
